Separate out `readAndProcessLines`
